+++ discord/member.py
@@ -44,6 +44,7 @@ from .flags import MemberFlags
 from .object import Object
 from .permissions import Permissions
 from .primary_guild import PrimaryGuild
+from .role import RoleColours
 from .user import BaseUser, User, _UserTag
 from .utils import MISSING
 
@@ -61,6 +62,7 @@ if TYPE_CHECKING:
     from .role import Role
     from .state import ConnectionState
     from .types.activity import PartialPresenceUpdate
+    from .types.member import Member
     from .types.member import Member as MemberPayload
     from .types.member import MemberWithUser as MemberWithUserPayload
     from .types.member import UserWithMember as UserWithMemberPayload
@@ -546,11 +548,33 @@ class Member(discord.abc.Messageable, _UserTag):
 
     @property
     def colour(self) -> Colour:
-        """A property that returns a colour denoting the rendered colour
+        """A property that returns a colour denoting the rendered primary colour
         for the member. If the default colour is the one rendered then an instance
         of :meth:`Colour.default` is returned.
 
-        There is an alias for this named :attr:`color`.
+        This is an alias for ``Member.colours.primary``.
+        """
+        return self.colours.primary
+
+    @property
+    def color(self) -> Colour:
+        """A property that returns a color denoting the primary rendered color for
+        the member. If the default color is the one rendered then an instance of :meth:`Colour.default`
+        is returned.
+
+        This is an alias for ``Member.colours.primary``.
+        """
+        return self.colours.primary
+
+    @property
+    def colours(self) -> RoleColours:
+        """A property that returns the rendered :class:`RoleColours` for
+        the member. If the default color is the one rendered then an instance of :meth:`RoleColours.default`
+        is returned.
+
+        There is an alias for this named :attr:`colors`.
+
+        .. versionadded:: 2.8
         """
 
         roles = self.roles[1:]  # remove @everyone
@@ -559,19 +583,21 @@ class Member(discord.abc.Messageable, _UserTag):
         # if the highest is the default colour then the next one with a colour
         # is chosen instead
         for role in reversed(roles):
-            if role.colour.value:
-                return role.colour
-        return Colour.default()
+            if role.colours.primary.value:
+                return role.colours
+        return RoleColours.default()
 
     @property
-    def color(self) -> Colour:
-        """A property that returns a color denoting the rendered color for
-        the member. If the default color is the one rendered then an instance of :meth:`Colour.default`
-        is returned.
+    def colors(self) -> RoleColours:
+        """A property that returns the rendered :class:`RoleColours` for the member.
+        If the default color is the one rendered then an instance
+        of :meth:`Colour.default` is returned.
 
-        There is an alias for this named :attr:`colour`.
+        This is an alias for :attr:`colours`.
+
+        .. versionadded:: 2.8
         """
-        return self.colour
+        return self.colours
 
     @property
     def roles(self) -> list[Role]: